requirements.txt
rdkit-pypi>=2023.9
pandas>=2.2
scikit-learn>=1.4
torch>=2.2
torch-geometric>=2.5
dgl>=2.0
optuna>=3.6
tensorboard>=2.16
mlflow>=2.12
pyyaml>=6.0
pytest>=8.0
ruff>=0.4
black>=24.3

